% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataObject.R, R/SystemMetadata.R
\docType{methods}
\name{addAccessRule,DataObject,data.frame-method}
\alias{addAccessRule}
\alias{addAccessRule,DataObject,data.frame-method}
\alias{addAccessRule,SystemMetadata,character-method}
\alias{addAccessRule,SystemMetadata,data.frame-method}
\title{Add a Rule to the AccessPolicy}
\usage{
\S4method{addAccessRule}{DataObject,data.frame}(x, y)

addAccessRule(x, y, ...)

\S4method{addAccessRule}{SystemMetadata,character}(x, y, permission)

\S4method{addAccessRule}{SystemMetadata,data.frame}(x, y)
}
\arguments{
\item{x}{The object instance to which to add the rules}

\item{y}{The subject of the rule to be added, or a data frame of subject/permission tuples}

\item{...}{(Not yet used)}

\item{permission}{The permission to be applied to subject if x is character (read, write, changePermission)}
}
\value{
DataObject with modified access rules
}
\description{
Add a Rule to the AccessPolicy

Add one or more access rules to a SystemMetadata object.
}
\details{
To be called prior to creating the object in DataONE.  When called before
creating the object, adds rules to the access policy for this object that
control access when uploaded to DataONE.  If called after creation, it
will only change the system metadata locally, and will not have any effect
on remotely uploaded copies of the DataObject.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataObject,y = data.frame}: 

\item \code{x = SystemMetadata,y = character}: 

\item \code{x = SystemMetadata,y = data.frame}: 
}}
\examples{
\dontrun{
accessRules <- data.frame(subject=c("uid=smith,ou=Account,dc=example,dc=com",
  "uid=slaughter,o=unaffiliated,dc=example,dc=org"), permission=c("write", "changePermission"))
sciObj <- addAccessRule(sciObj, accessRules)
}
\dontrun{
sysmeta <- addAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "write")
accessRules <- data.frame(subject=c("uid=smith,ou=Account,dc=example,dc=com",
  "uid=slaughter,o=unaffiliated,dc=example,dc=org"), permission=c("write", "changePermission"))
sysmeta <- addAccessRule(sysmeta, accessRules)
}
}
\seealso{
\code{\link[=SystemMetadata-class]{SystemMetadata}}{ class description.}

\code{\link[=DataObject-class]{DataObject}}{ class description.}
}

