% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataObject.R
\name{addAccessRules}
\alias{addAccessRules}
\title{Add a Rule to the AccessPolicy}
\usage{
addAccessRules(x, accessRules)
}
\arguments{
\item{x}{DataObject}

\item{accessRules}{A data frame containing one or more access rules}
}
\value{
DataObject with modified access rules
}
\description{
To be called prior to creating the object in DataONE.  When called before
creating the object, adds a rule to the access policy that makes this object
publicly readable.  If called after creation, it will only change the system
metadata locally, and will not have any affect.
}
\examples{
\dontrun{
accessRules <- data.frame(subject=c("uid=smith,ou=Account,dc=example,dc=com", "uid=slaughter,o=unaffiliated,dc=example,dc=org"), permission=c("write", "changePermission"))
sciObj <- addAccessRules(sciObj, accessRules)
}
}

