% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{addData}
\alias{addData}
\alias{addData,DataPackage,DataObject-method}
\title{Add a DataObject to the DataPackage}
\usage{
addData(x, do, ...)

\S4method{addData}{DataPackage,DataObject}(x, do)
}
\arguments{
\item{x}{A DataPackage instance}

\item{do}{A DataObject identifier}

\item{...}{(not yet used)}
}
\description{
The DataObject is added to the DataPackage, making it available for
retrieval and eventual upload using the method \code{\link[dataone]{uploadDataPackage}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataPackage,do = DataObject}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

