% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\name{calculateChecksum}
\alias{calculateChecksum}
\alias{calculateChecksum,DataObject-method}
\title{Calculate a checksum for the DataObject using the specified checksum algorithm}
\usage{
calculateChecksum(x, ...)

\S4method{calculateChecksum}{DataObject}(x, checksumAlgorithm = "SHA256", ...)
}
\arguments{
\item{x}{A DataObject instance}

\item{...}{Additional parameters (not yet used)}

\item{checksumAlgorithm}{a \code{character} value specifying the checksum algorithm to use (i.e "MD5" or "SHA1" or "SHA256")}
}
\value{
The calculated checksum
}
\description{
calculates a checksum
}
\note{
this method is intended for internal package use only.
}
