% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{canRead}
\alias{canRead}
\alias{canRead,DataObject,character-method}
\title{Test whether the provided subject can read an object.}
\usage{
canRead(x, subject, ...)

\S4method{canRead}{DataObject,character}(x, subject)
}
\arguments{
\item{x}{DataObject}

\item{subject}{: the subject name of the person/system to check for read permissions}

\item{...}{Additional arguments}
}
\value{
TRUE or FALSE
}
\description{
Using the AccessPolicy, tests whether the subject has read permission
for the object.  This method is meant work prior to submission, so uses
only the AccessPolicy to determine who can read (Not the rightsHolder field,
which always can read.)
}
\details{
The subject name used in both the AccessPolicy and in the \code{'subject'}
argument to this method is a string value, but is generally formatted as an X.509
name formatted according to RFC 2253.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataObject,subject = character}: 
}}

