% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataObject.R
\name{DataObject}
\alias{DataObject}
\alias{construct-DataObject}
\title{Create a DataObject object}
\usage{
DataObject(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
a DataObject
}
\description{
Create a DataObject object
}
\seealso{
\code{\link[=DataObject-class]{DataObject}}{ class description.}
}

