% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{DataPackage}
\alias{DataPackage}
\alias{DataPackage,ANY-method}
\alias{construct-DataPackage}
\title{Create a DataPackage object}
\usage{
DataPackage(...)

\S4method{DataPackage}{ANY}()
}
\arguments{
\item{...}{(Not yet used)}
}
\description{
The DataPackage() method is a shortcut to creating a DataPackage object, as does
not allow specifying any options that the \code{\link[=initialize-DataPackage]{initialize}} method allows.
}
\examples{
pkg <- DataPackage()
}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

