% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{SystemMetadata,XMLInternalElementNode-method}
\alias{SystemMetadata,XMLInternalElementNode-method}
\alias{construct-SystemMetadata-XMLInternalElementNode}
\title{Create a SystemMetadata from an XML document}
\usage{
\S4method{SystemMetadata}{XMLInternalElementNode}(sysmeta, ...)
}
\arguments{
\item{sysmeta}{A value of type \code{"XMLInternalElementNode"}, containing the parsed XML element with SystemMetadata fields.}

\item{...}{(Not implemented ye)}
}
\description{
Construct a new SystemMetadata instance by using the fields from an XML representation of the
SystemMetadata.
}

