% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{SystemMetadata}
\alias{SystemMetadata}
\alias{SystemMetadata,ANY-method}
\alias{construct-SystemMetadata}
\title{Create DataONE SystemMetadata object}
\usage{
SystemMetadata(...)

\S4method{SystemMetadata}{ANY}(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
the SystemMetadata object
}
\description{
A class representing DataONE SystemMetadata, which is core information about objects stored in a repository
and needed to manage those objects across systems.  SystemMetadata contains basic identification, ownership,
access policy, replication policy, and related metadata.
}
\seealso{
\code{\link[=SystemMetadata-class]{SystemMetadata}}{ class description.}
}

