% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{containsId}
\alias{containsId}
\title{Returns true if the specified object is a member of the package}
\usage{
containsId(x, identifier, ...)
}
\description{
Returns true if the specified object is a member of the package
}

