% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\docType{methods}
\name{createFromTriples}
\alias{createFromTriples}
\alias{createFromTriples,ResourceMap,data.frame,character-method}
\title{Populate a ResourceMap with RDF relationships from data.frame.}
\usage{
createFromTriples(.Object, relations, identifiers)

\S4method{createFromTriples}{ResourceMap,data.frame,character}(.Object,
  relations, identifiers)
}
\arguments{
\item{.Object}{a ResourceMap}

\item{relations}{A data.frame to read relationships from}

\item{identifiers}{A list of the identifiers of data objects cotained in the associated data package}
}
\description{
RDF relationships are added to a ResourceMap object from a data.frame that
contains RDF triples. For example, relationships can be exported from a DataPackage via
\code{\link{getRelationships}}. The resulting data.frame is then read by \code{createFromTriples}
to create the ResourceMap.
}
\details{
The \code{identifiers} parameter contains the identifiers of all data objects in the DataPackage.
For each data objects, additional relationships will be added that are required by the OAI-ORE specification,
for example a Dublin Core identifier statement is added.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = ResourceMap,relations = data.frame,identifiers = character}: 
}}
\seealso{
\code{\link[=ResourceMap-class]{ResourceMap}}{ class description.}
}

