% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DataObject-class}
\alias{DataObject,}
\alias{DataObject-class}
\title{DataObject is a wrapper for raw data to include important system-level metadata about the object.}
\description{
DataObject is a wrapper class that associates raw data with system-level metadata
describing the object.  The system metadata includes attributes such as the object's identifier,
type, size, owner, version relationship to other objects, access rules, and other critical metadata.
The SystemMetadata is compliant with the DataONE federated network's definition of SystemMetadata, and
is encapsulated as a separate object that can be manipulated as needed.
}
\details{
A DataObject can be constructed by passing the data and system metadata to the new() method, or by passing
an identifier, data, format, user, and DataONE node identifier, in which case a SystemMetadata instance will
be generated with these fields and others that are calculated (such as size and checksum).

An obvious limitation of the current implementation is that all data are stored in memory, which limits
the effective size of objects.  It would be useful to add a data cache that stores larger objects on the file
system.
}
\section{Slots}{

\describe{
\item{\code{sysmeta}}{value of type \code{"SystemMetadata"}, containing the metadata about the object}

\item{\code{data}}{value of type \code{"raw"}, containing the data represented in this object}
}}
\examples{
do <- new("DataObject", "id1", charToRaw("1,2,3\\n4,5,6\\n"), "text/csv", "matt", "urn:node:KNB")
id <- getIdentifier(do)
fmt <- getFormatId(do)
}
\author{
Matthew Jones
}
\keyword{classes}

