% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataObject.R
\name{DataObject}
\alias{DataObject}
\title{Construct a DataObject instance.}
\usage{
DataObject(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
a DataObject
}
\description{
Construct a DataObject instance.
}

