% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{datapack-deprecated}
\alias{datapack-deprecated}
\title{Deprecated Methods}
\description{
The following items are deprecated in this release of datapack and will be
marked as Defunct and removed in a future version.
}
\section{These methods are deprecated}{

\itemize{
 \item{\code{\link{recordDerivation}}}{: Record derivation relationships between objects in a DataPackage.}
 \item{\code{\link{addData}}}{: Add a DataObject to the DataPackage.}
}
}

\keyword{internal}
