% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\docType{methods}
\name{freeResourceMap}
\alias{freeResourceMap}
\alias{freeResourceMap,ResourceMap-method}
\title{Free memory used by a ResouceMap.}
\usage{
freeResourceMap(.Object)

\S4method{freeResourceMap}{ResourceMap}(.Object)
}
\arguments{
\item{.Object}{a ResourceMap}
}
\description{
The resources allocated by the redland RDF package are freed. The ResourceMap
object should be deleted immediately following this call.
}
\section{Methods (by class)}{
\itemize{
\item \code{ResourceMap}: 
}}
\seealso{
\code{\link[=ResourceMap-class]{ResourceMap}}{ class description.}
}

