% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R, R/DataPackage.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,DataObject,ANY-method}
\alias{getData,DataPackage,character-method}
\title{Get the data content of a specified data object}
\usage{
getData(x, id = NA, ...)

\S4method{getData}{DataObject,ANY}(x)

\S4method{getData}{DataPackage,character}(x, id)
}
\arguments{
\item{x}{DataObject or DataPackage: the data structure from where to get the data}

\item{id}{Missing or character: if \code{'x'} is DataPackage, the identifier of the
package member to get data from}

\item{...}{Additional arguments}
}
\value{
raw representation of the data
}
\description{
Get the data content of a specified data object
}
\seealso{
\code{\link[=DataObject-class]{DataObject}}{ class description.}
}

