% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getIdentifier}
\alias{getIdentifier}
\alias{getIdentifier,DataObject-method}
\title{Get the Identifier of the DataObject}
\usage{
getIdentifier(x, ...)

\S4method{getIdentifier}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\value{
the identifier
}
\description{
Get the Identifier of the DataObject
}
\section{Methods (by class)}{
\itemize{
\item \code{DataObject}: 
}}

