% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getIdentifiers}
\alias{getIdentifiers}
\alias{getIdentifiers,DataPackage-method}
\title{Get the Identifiers of Package Members}
\usage{
getIdentifiers(x, ...)

\S4method{getIdentifiers}{DataPackage}(x)
}
\arguments{
\item{x}{A DataPackage instance}

\item{...}{(not yet used)}
}
\value{
A list of identifiers
}
\description{
The identifiers of the objects in the package are retrieved and returned as a list.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataPackage}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

