% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getMember}
\alias{getMember}
\alias{getMember,DataPackage,character-method}
\title{Return the Package Member by Identifier}
\usage{
getMember(x, identifier, ...)

\S4method{getMember}{DataPackage,character}(x, identifier)
}
\arguments{
\item{x}{A DataPackage object}

\item{identifier}{A DataObject identifier}

\item{...}{(Not yet used)}
}
\value{
A DataObject if the member is found, or NULL if not
}
\description{
Given the identifier of a member of the data package, return the DataObject
representation of the member.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataPackage,identifier = character}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

