% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRelationships}
\alias{getRelationships}
\title{Return relationships between package objects}
\usage{
getRelationships(x, ...)
}
\description{
Relationships between objects in a package are defined using the \code{'insertRelationship'} call and retrieved
using \code{getRetaionships}. These relationships are returned in a data frame with \code{'subject'}, \code{'predicate'}, \code{'objects'}
as the columns, ordered by "subject"
}

