% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceMap.R
\name{getTriples}
\alias{getTriples}
\title{Parse an RDF/XML resource map from a file}
\usage{
getTriples(x, ...)
}
\arguments{
\item{x}{ResourceMap}

\item{file}{a file containing a resource map in RDF/XML that will be parsed into the ResourceMap object}
}
\value{
x the ResourceMap containing the parsed RDF/XML content
}
\description{
parseRDF reads a file containing an RDF model in RDF/XML format and initializes
a ResourceMap based on this content
}
\details{
This method resets the slot ResourceMap@world so any previously stored triples are discarded, allowing
for a clean model object in which to parse the new RDF content into. It is assumed that the content is a
valid ORE resource map and no validation checks specific to the OAI-ORE content model are performed.
}
