% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,DataObject-method}
\alias{initialize,DataObject-method}
\title{Intialize a DataObject instance with values passed to the constructor function.
When initializing a DataObject using passed in data, one can either pass
in the \code{'id'} param as a \code{'SystemMetadata'} object, or as a \code{'character'} string
representing the identifier for an object along with parameters for format, user,and associated member node.
In either case, the \code{'data'} param holds the \code{'raw'} data.}
\usage{
\S4method{initialize}{DataObject}(.Object, id, data, format = NA, user = NA,
  mnNodeId = NA)
}
\description{
Intialize a DataObject instance with values passed to the constructor function.
When initializing a DataObject using passed in data, one can either pass
in the \code{'id'} param as a \code{'SystemMetadata'} object, or as a \code{'character'} string
representing the identifier for an object along with parameters for format, user,and associated member node.
In either case, the \code{'data'} param holds the \code{'raw'} data.
}

