% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\docType{methods}
\name{initialize,ResourceMap-method}
\alias{initialize,ResourceMap-method}
\title{Initialize a OAI-ORE ResourceMap object.}
\usage{
\S4method{initialize}{ResourceMap}(.Object, id = as.character(NA))
}
\arguments{
\item{.Object}{a ResourceMap object}

\item{id}{a unique identifier to identify this ResourceMap. This id will be used internally in the ResourceMap.}
}
\value{
the ResourceMap object
}
\description{
Create the objects used by the ResourceMap object to store and process
relationships (RDF triples) of the DataPackage.
}

