% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{insertDerivation}
\alias{insertDerivation}
\alias{insertDerivation,DataPackage-method}
\title{Add provenance relationships to a DataPackage for a script execution.}
\usage{
insertDerivation(x, ...)

\S4method{insertDerivation}{DataPackage}(x, sources = list(),
  program = as.character(NA), derivations = list(), ...)
}
\arguments{
\item{x}{The \code{DataPackage} to add provenance relationships to.}

\item{...}{Additional parameters}

\item{sources}{A list of DataObjects for files that were read by the program. Alternatively, a list 
of DataObject identifiers can be specified.}

\item{program}{The DataObject created for the R script. Alternatively the DataObject identifier can
be specified.}

\item{derivations}{A list of DataObjects for files that were generated by the program. Alternatively, a list 
of DataObject identifiers can be specified.}
}
\description{
A DataPackage can store provenance relationships between the DataObjects in the package. 
The method \code{insertDerivation} will add provenance relationships between an R script, the files that
it used and the files that it generated.
}
\details{
This method operates on a DataPackage that has had DataObjects for the script, inputs and outputs
previously added to it. Provenance relationships from the ProvONE data model are added to the package. The
ProvONE data model can be viewed at \url{https://purl.dataone.org/provone-v1-dev}.
}
\examples{
library(datapack)
dp <- new("DataPackage")
# Add the script to the DataPackage
progFile <- system.file("./extdata/pkg-example/logit-regression-example.R", package="datapack")
progObj <- new("DataObject", format="application/R", filename=progFile)
dp <- addData(dp, progObj)

# Add a script input to the DataPackage
inFile <- system.file("./extdata/pkg-example/binary.csv", package="datapack") 
inObj <- new("DataObject", format="text/csv", filename=inFile)
dp <- addData(dp, inObj)

# Add a script output to the DataPackage
outFile <- system.file("./extdata/pkg-example/gre-predicted.png", package="datapack")
outObj <- new("DataObject", format="image/png", file=outFile)
dp <- addData(dp, outObj)

# Add the provenenace relationshps, linking the input and output to the script execution
# Note: 'sources' and 'derivations' can also be lists of "DataObjects" or "DataObject' identifiers
dp <- insertDerivation(dp, sources = inObj, program = progObj, derivations = outObj) 
# View the results
head(getRelationships(dp))
}

