% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insertRelationship}
\alias{insertRelationship}
\title{Record relationships between objects in a DataPackage}
\usage{
insertRelationship(x, subjectID, objectIDs, predicate, ...)
}
\arguments{
\item{x}{a DataPackage object}

\item{subjectID}{the identifier of the object that will be the subject of the relationship}

\item{predicate}{the IRI of the predicate of the relationship}

\item{objectIDS}{a list of identifiers of the object of the relationship (a relationship is recorded for each objectID)}
}
\description{
Record a relationship of the form "subject -> predicate -> object", as defined by the Resource Description Framework (RDF), i.e.
an RDF triple. For use with DataONE, a best practice is to specifiy the subject and predicate as DataONE persistent identifiers
(https://mule1.dataone.org/ArchitectureDocs-current/design/PIDs.html). If the objects are not known to DataONE, then local identifiers can be
used, and these local identifiers may be promoted to DataONE PIDs when the package is uploaded to a DataONE member node.
The predicate is typically an RDF property (as a IRI) from a schema supported by DataONE, i.e. "http://www.w3.org/ns/prov#wasGeneratedBy"
}
\details{
A relationship is created for each value in the list "objectIDs". See examples for more information.
#'
}
\examples{
\dontrun{
dp <- DataPackage()
insertRelationship(dp, "/Users/smith/scripts/genFields.R",
                       "http://www.w3.org/ns/prov#used",
                       "https://knb.ecoinformatics.org/knb/d1/mn/v1/object/doi:1234/_030MXTI009R00_20030812.40.1")

}
}

