% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{parseSystemMetadata}
\alias{parseSystemMetadata}
\alias{parseSystemMetadata,SystemMetadata,XMLInternalElementNode-method}
\alias{parseSystemMetadata,SystemMetadata-method}
\title{parse metadata}
\usage{
parseSystemMetadata(sysmeta, xml, ...)

\S4method{parseSystemMetadata}{SystemMetadata,XMLInternalElementNode}(sysmeta,
  xml, ...)
}
\arguments{
\item{sysmeta}{the \code{SystemMetadata} object}

\item{xml}{the XML representation of the capabilities, as an XMLInternalElementNode}

\item{...}{additional arguments passed to other functions or methods}
}
\value{
the SystemMetadata object representing an object
}
\description{
Parse an XML representation of system metadata, and set the object slots with obtained values.
}
\author{
jones
}

