% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{plotRelationships}
\alias{plotRelationships}
\alias{plotRelationships,DataPackage-method}
\title{Plot derivation relationships obtained from getRelationships}
\usage{
plotRelationships(x, ...)

\S4method{plotRelationships}{DataPackage}(x, col = NULL, ...)
}
\arguments{
\item{x}{a DataPackage object}

\item{...}{other options passed to the igraph plot function}

\item{col}{vector of colors used for plotting}
}
\description{
Creates graph of dataPackage object generated from getRelationships
}
\seealso{
\code{\link{DataPackage-class}}
}
