% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{removeAccessRule}
\alias{removeAccessRule}
\alias{removeAccessRule,SystemMetadata-method}
\title{Remove an access rule from the specified object.}
\usage{
removeAccessRule(x, ...)

\S4method{removeAccessRule}{SystemMetadata}(x, y, ...)
}
\arguments{
\item{x}{The object instance to which to remove the rule}

\item{...}{Additional arguments
\itemize{
  \item{permission The permission to be remove to subject if x is character (read, write, changePermission)}
}}

\item{y}{The subject of the rule to be removed}
}
\value{
the SystemMetadata object with the updated access policy.
}
\description{
Remove access rules from the access policy of the specified object.
}
\examples{
# Parameter "y" can be character string containing the subject of the access rule:
sysmeta <- new("SystemMetadata")
sysmeta <- addAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "write")
sysmeta <- addAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "changePermission")
sysmeta <- removeAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "changePermission")

# Alternatively, parameter "y" can be a data.frame containing one or more access rules:
sysmeta <- addAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "write")
accessRules <- data.frame(subject=c("uid=smith,ou=Account,dc=example,dc=com", 
  "uid=slaughter,o=unaffiliated,dc=example,dc=org"), permission=c("write", "changePermission"))
sysmeta <- removeAccessRule(sysmeta, accessRules)
}
\seealso{
\code{\link{SystemMetadata-class}}
}
