% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{removeMember}
\alias{removeMember}
\alias{removeMember,DataPackage,character-method}
\title{Remove the Specified Member from the Package}
\usage{
removeMember(x, identifier, ...)

\S4method{removeMember}{DataPackage,character}(x, identifier)
}
\arguments{
\item{x}{a Datapackage object}

\item{identifier}{an identifier for a DataObject}

\item{...}{(Not yet used)}
}
\description{
Given the identifier of a member of the data package, delete the DataObject
representation of the member.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataPackage,identifier = character}: 
}}
\examples{
dp <- new("DataPackage")
data <- charToRaw("1,2,3\\n4,5,6")
do <- new("DataObject", id="myNewId", dataobj=data, format="text/csv", user="jsmith")
addData(dp, do)
removeMember(dp, "myNewId")
}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

