% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{removeRelationships}
\alias{removeRelationships}
\alias{removeRelationships,DataPackage-method}
\title{Remove package relationships that match search terms.}
\usage{
removeRelationships(x, ...)

\S4method{removeRelationships}{DataPackage}(x, subject = as.character(NA),
  predicate = as.character(NA), object = as.character(NA), ...)
}
\arguments{
\item{x}{A DataPackage object}

\item{...}{(Not yet used)}

\item{subject}{A character value for subject to match in the package relationships}

\item{predicate}{A character value for predicate to match in the package relationships}

\item{object}{A character value for object to match in the package relationships}
}
\description{
When package members are updated, they receive a new identifier. It is therefor
necessary to update the package relationships to update occurances of the old identifier
with the new one when the old identifier appears in the "subject" or "object" of a 
relationship.
}
\seealso{
\code{\link{DataPackage-class}}
}
