% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{selectMember}
\alias{selectMember}
\alias{selectMember,DataPackage-method}
\title{Return identifiers for objects that match search criteria}
\usage{
selectMember(x, ...)

\S4method{selectMember}{DataPackage}(x, name, value)
}
\arguments{
\item{x}{A DataPackage instance}

\item{...}{(Not yet used)}

\item{identifier}{A DataObject identifier}
}
\value{
A DataObject if the member is found, or NULL if not
}
\description{
Given the identifier of a member of the data package, return the DataObject
}
\examples{
dp 

<- new("DataPackage")
data <- charToRaw("1,2,3\\n4,5,6")
do <- new("DataObject", id="myNewId", dataobj=data, format="text/csv", user="jsmith")
dp <- addData(dp, do)
do2 <- getMember(dp, "myNewId")
}
\seealso{
\code{\link{DataPackage-class}}
}
