% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{serializeSystemMetadata}
\alias{serialize,SystemMetadata-method}
\alias{serializeSystemMetadata}
\alias{serializeSystemMetadata,SystemMetadata-method}
\title{serialize system metadata}
\usage{
serializeSystemMetadata(sysmeta, ...)

\S4method{serializeSystemMetadata}{SystemMetadata}(sysmeta)
}
\arguments{
\item{sysmeta}{the SystemMetadata instance to be serialized}
}
\value{
the character string representing a SystemMetadata object
}
\description{
Serialize an XML representation of system metadata.
}
\author{
jones
}

