% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{serializePackage}
\alias{serializePackage}
\title{Serialize the DataPackage object}
\usage{
serializePackage(.Object, file, ...)
}
\arguments{
\item{.Object}{the DataPackage object}

\item{file}{the file to which the ResourceMap will be serialized}

\item{id}{a unique identifier for the serialization. If a value is not specified, one will be generated}

\item{syntaxName}{name of the syntax to use for serialization - default is "rdfxml"}

\item{mimetype}{the mimetype of the serialized output - the default is "application/rdf+xml"}

\item{namespaces}{a data frame containing one or more namespaces and their associated prefix}

\item{syntaxURI}{URI of the serialization syntax}
}
\description{
Datapackage relationships are serialized as a OAI-ORE ResourceMap
}

