% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{serializeToBagit}
\alias{serializeToBagit}
\alias{serializeToBagit,DataPackage-method}
\title{Serialize A DataPackage into a Bagit Archive File}
\usage{
serializeToBagit(.Object, ...)

\S4method{serializeToBagit}{DataPackage}(.Object)
}
\arguments{
\item{.Object}{A DataPackage object}

\item{...}{Additional arguments}
}
\value{
A zip file containing a Bagit archive.
}
\description{
The Bagit packaging format \url{https://tools.ietf.org/html/draft-kunze-bagit-08}
is used to prepare an archive file that contains the contents of a DataPackage.
}
\details{
A Bagit Archive File is created by copying each member of a DataPackge, and preparing
files that describe the files in the archive, including information about the size of the files
and a checksum for each file. These metadata files and the data files are then packaged into
a single zip file.
}
\examples{
# Create the first data object
dp <- new("DataPackage")
data <- charToRaw("1,2,3\\n4,5,6")
do <- new("DataObject", id="do1", dataobj=data, format="text/csv", user="jsmith")
addData(dp, do)
# Create a second data object
data2 <- charToRaw("7,8,9\\n4,10,11")
do2 <- new("DataObject", id="do2", dataobj=data2, format="text/csv", user="jsmith")
addData(dp, do2)
# Create a relationship between the two data objects
recordDerivation(dp, "do2", "do2")
# Write out the data package to a Bagit file
bagitFile <- serializeToBagit(dp)
}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

