% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{SystemMetadata,XMLInternalElementNode-method}
\alias{SystemMetadata,XMLInternalElementNode-method}
\title{Construct a SystemMetadata object, with all fields set to the value found in an XML document}
\usage{
\S4method{SystemMetadata}{XMLInternalElementNode}(sysmeta)
}
\arguments{
\item{sysmeta}{value of type \code{"XMLInternalElementNode"}, containing the parsed XML element with SystemMetadata fields.}
}
\value{
the SystemMetadata object representing an object
}
\description{
Construct a new SystemMetadata instance by using the fields from an XML representation of the
SystemMetadata.  The XML should be
}
\author{
jones
}

