% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{SystemMetadata}
\alias{SystemMetadata}
\alias{SystemMetadata,XMLInternalElementNode-method}
\title{Create DataONE SystemMetadata object}
\usage{
SystemMetadata(...)

\S4method{SystemMetadata}{XMLInternalElementNode}(x, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{x}{A value of type \code{"XMLInternalElementNode"}, containing the parsed XML element with SystemMetadata fields.}
}
\description{
A class representing DataONE SystemMetadata, which is core information about objects stored in a repository
and needed to manage those objects across systems.  SystemMetadata contains basic identification, ownership,
access policy, replication policy, and related metadata.

If the *sysmeta* parameter is specified, then construct a new SystemMetadata instance by using the fields from 
an XML representation of the SystemMetadata.
}
\seealso{
\code{\link{SystemMetadata-class}}
}
