% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{updateMetadata}
\alias{updateMetadata}
\alias{updateMetadata,DataPackage-method}
\title{Update selected elements of the XML content of a DataOBject in a DataPackage (aka package member).}
\usage{
updateMetadata(x, do, ...)

\S4method{updateMetadata}{DataPackage}(x, do, xpath, replacement,
  newId = as.character(NA), ...)
}
\arguments{
\item{x}{a DataPackage instance}

\item{do}{A DataObject instance}

\item{...}{(Not yet used)}

\item{xpath}{A \code{character} value specifying the location in the XML to update.}

\item{replacement}{A \code{character} value that will replace the elements found with the \code{xpath}.}

\item{newId}{A value of type \code{"character"} which will replace the identifier for this DataObject.}
}
\examples{
# Create a DataObject and add it to the DataPackage
dp <- new("DataPackage")

}
\seealso{
\code{\link{DataPackage-class}}
}
