% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{updateRelationships}
\alias{updateRelationships}
\alias{updateRelationships,DataPackage-method}
\title{Update package relationships by replacing an old identifier with a new one.}
\usage{
updateRelationships(x, ...)

\S4method{updateRelationships}{DataPackage}(x, id, newId, ...)
}
\arguments{
\item{x}{A DataPackage object}

\item{...}{(Not yet used)}

\item{id}{A character value containing the identifier to be replaced.}

\item{newId}{A character value containing the identifier that will replace the old identifier.}
}
\description{
When package members are updated, they receive a new identifier (replaceMember). It is therefor
necessary to update the package relationships to update occurrences of the old identifier
with the new one when the old identifier appears in the "subject" or "object" of a 
relationship.
}
\seealso{
\code{\link{DataPackage-class}}
}
