% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{validate}
\alias{validate}
\alias{validate,SystemMetadata-method}
\title{Validate a SystemMetadata object.}
\usage{
validate(object, ...)

\S4method{validate}{SystemMetadata}(object, ...)
}
\arguments{
\item{object}{the instance to be validated}

\item{...}{(Additional parameters)}
}
\value{
logical, \code{TRUE} if the SystemMetadata object is valid, else a list of strings detailing errors
}
\description{
Validate a system metadata object, ensuring that required fields are present and of the right type.
}
\section{Methods (by class)}{
\itemize{
\item \code{SystemMetadata}: 
}}
\seealso{
\code{\link[=SystemMetadata-class]{SystemMetadata}}{ class description.}
}

