% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addData}
\alias{addData}
\title{Add a DataObject to the DataPackage}
\usage{
addData(x, do, ...)
}
\arguments{
\item{x}{: DataPackage}

\item{...}{: (not yet used)}

\item{d1object}{: DataObject}
}
\description{
Includes the DataObject in the DataPackage data Map, making it available for
retrieval and eventual upload (via createPackage).
}
\author{
rnahf
}

