% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{addData}
\alias{addData}
\title{Add a DataObject to the DataPackage}
\usage{
addData(x, do, ...)
}
\arguments{
\item{x}{: DataPackage}

\item{do}{: DataObject, or identifier of an object on the DataONE network}

\item{...}{: (not yet used)}
}
\description{
Includes the DataObject in the DataPackage data Map, making it available for
retrieval and eventual upload (via createPackage).
}

