% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{clearAccessPolicy}
\alias{clearAccessPolicy}
\alias{clearAccessPolicy,SystemMetadata-method}
\title{Clear the accessPolicy from the specified object.}
\usage{
clearAccessPolicy(x, ...)

\S4method{clearAccessPolicy}{SystemMetadata}(x, ...)
}
\arguments{
\item{x}{the instance to clear access rules from.}

\item{...}{(Additional parameters)}
}
\value{
The SystemMetadata object with the cleared access policy.
}
\description{
Clears the accessPolicy from the specified object by overwriting
all existing access rules set on the object.
}
\examples{
sysmeta <- new("SystemMetadata")
sysmeta <- addAccessRule(sysmeta, "uid=smith,ou=Account,dc=example,dc=com", "write")
sysmeta <- clearAccessPolicy(sysmeta)
}
\seealso{
\code{\link{SystemMetadata-class}}
}

