% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{containsId}
\alias{containsId}
\alias{containsId,DataPackage,character-method}
\title{Returns true if the specified object is a member of the package}
\usage{
containsId(x, identifier, ...)

\S4method{containsId}{DataPackage,character}(x, identifier)
}
\arguments{
\item{x}{A DataPackage object}

\item{identifier}{The DataObject identifier to check for inclusion in the DataPackage}

\item{...}{(Not yet used)}
}
\value{
A logical - a value of TRUE indicates that the DataObject is in the DataPackage
}
\description{
Returns true if the specified object is a member of the package
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataPackage,identifier = character}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

