% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\name{createFromTriples}
\alias{createFromTriples}
\title{Populate a ResourceMap with RDF relationships.}
\usage{
createFromTriples(.Object, relations, identifiers)
}
\arguments{
\item{.Object}{a ResourceMap}

\item{file}{the file to which the ResourceMap will be serialized.}
}
\description{
RDF relationships are added to a ResourceMap object from a data.frame that
contains RDF triples.
}

