% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dmsg}
\alias{dmsg}
\title{Print a debugging message to stderr.}
\usage{
dmsg(msg)
}
\arguments{
\item{msg}{the message to be printed}
}
\description{
Print a debugging message to stderr.
}
\details{
Only print the message if the option "datapackage.debugging_mode" is TRUE.
}

