% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\name{freeResourceMap}
\alias{freeResourceMap}
\title{Free memory used by a ResouceMap.}
\usage{
freeResourceMap(.Object)
}
\arguments{
\item{.Object}{a ResourceMap}
}
\description{
The resources allocated by the redland RDF package are freed. The ResourceMap
object should be deleted immediately following this call.
}

