% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceMap.R
\name{freeResourceMap}
\alias{freeResourceMap}
\alias{freeResourceMap,ResourceMap-method}
\title{Free memory used by a ResouceMap}
\usage{
freeResourceMap(x)

\S4method{freeResourceMap}{ResourceMap}(x)
}
\arguments{
\item{x}{a ResourceMap}
}
\description{
The resources allocated by the redland RDF package are freed. The ResourceMap
object should be deleted immediately following this call.
}
\seealso{
\code{\link{ResourceMap-class}}
}
