% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getData,DataPackage,character-method}
\alias{getData,DataPackage,character-method}
\alias{getData,DataPackage-methods}
\title{Get the data content of a specified data object}
\usage{
\S4method{getData}{DataPackage,character}(x, id)
}
\arguments{
\item{x}{DataPackage: the data structure from where to get the data}

\item{id}{if \code{'x'} is DataPackage, the identifier of the package member to get data from}
}
\value{
raw representation of the data
}
\description{
Get the data content of a specified data object
}

