% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,DataObject,ANY-method}
\title{Get the data content of a specified data object}
\usage{
getData(x, id = NA, ...)

\S4method{getData}{DataObject,ANY}(x)
}
\arguments{
\item{x}{DataObject or DataPackage: the data structure from where to get the data}

\item{id}{Missing or character: if \code{'x'} is DataPackage, the identifier of the
package member to get data from}

\item{...}{Additional arguments}
}
\value{
raw representation of the data
}
\description{
Get the data content of a specified data object
}
\section{Methods (by class)}{
\itemize{
\item \code{x = DataObject,id = ANY}: 
}}

