% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\docType{methods}
\name{getFormatId}
\alias{getFormatId}
\alias{getFormatId,DataObject-method}
\title{Get the FormatId of the DataObject}
\usage{
getFormatId(x, ...)

\S4method{getFormatId}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\value{
the formatId
}
\description{
Get the FormatId of the DataObject
}
\examples{
data <- charToRaw("1,2,3\\n4,5,6\\n")
do <- new("DataObject", "id1", dataobj=data, "text/csv", 
  "uid=jones,DC=example,DC=com", "urn:node:KNB")
fmtId <- getFormatId(do)
}
\seealso{
\code{\link{DataObject-class}}
}

