% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataObject.R
\docType{methods}
\name{getFormatId}
\alias{getFormatId}
\alias{getFormatId,DataObject-method}
\title{Get the FormatId of the DataObject}
\usage{
getFormatId(x, ...)

\S4method{getFormatId}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\value{
the formatId
}
\description{
Get the FormatId of the DataObject
}
\section{Methods (by class)}{
\itemize{
\item \code{DataObject}: 
}}
\seealso{
\code{\link[=DataObject-class]{DataObject}}{ class description.}
}

