% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\docType{methods}
\name{getIdentifier}
\alias{getIdentifier}
\alias{getIdentifier,DataObject-method}
\title{Get the Identifier of the DataObject}
\usage{
getIdentifier(x, ...)

\S4method{getIdentifier}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\value{
the identifier
}
\description{
Get the Identifier of the DataObject
}
\examples{
data <- charToRaw("1,2,3\\n4,5,6\\n")
do <- new("DataObject", "id1", dataobj=data, "text/csv", 
  "uid=jones,DC=example,DC=com", "urn:node:KNB")
id <- getIdentifier(do)
}
\seealso{
\code{\link{DataObject-class}}
}

