% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\docType{methods}
\name{getIdentifier}
\alias{getIdentifier}
\alias{getIdentifier,DataObject-method}
\title{Get the Identifier of the DataObject}
\usage{
getIdentifier(x, ...)

\S4method{getIdentifier}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\value{
the identifier
}
\description{
Get the Identifier of the DataObject
}
\seealso{
\code{\link[=DataObject-class]{DataObject}}{ class description.}
}

