% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIdentifiers}
\alias{getIdentifiers}
\title{Get the Identifiers of Package Members}
\usage{
getIdentifiers(x, ...)
}
\arguments{
\item{x}{: DataPackage}

\item{...}{(not yet used)}
}
\value{
list of identifiers
}
\description{
Return the identifiers of the package members, as defined by the ResourceMap
}
\author{
rnahf
}

