% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{getIdentifiers}
\alias{getIdentifiers}
\title{Get the Identifiers of Package Members}
\usage{
getIdentifiers(x, ...)
}
\arguments{
\item{x}{: DataPackage}

\item{...}{(not yet used)}
}
\value{
list of identifiers
}
\description{
Return the identifiers of the package members, as defined by the ResourceMap
}

