% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getIdentifiers}
\alias{getIdentifiers}
\alias{getIdentifiers,DataPackage-method}
\title{Get the Identifiers of Package Members}
\usage{
getIdentifiers(x, ...)

\S4method{getIdentifiers}{DataPackage}(x)
}
\arguments{
\item{x}{A DataPackage instance}

\item{...}{(not yet used)}
}
\value{
A list of identifiers
}
\description{
The identifiers of the objects in the package are retrieved and returned as a list.
}
\examples{
dp <- new("DataPackage")
data <- charToRaw("1,2,3\\n4,5,6")
do <- new("DataObject", dataobj=data, format="text/csv", user="jsmith")
dp <- addMember(dp, do)
getIdentifiers(dp)
}
\seealso{
\code{\link{DataPackage-class}}
}
