% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getRelationships}
\alias{getRelationships}
\alias{getRelationships,DataPackage-method}
\title{Retrieve relationships of package objects}
\usage{
getRelationships(x, ...)

\S4method{getRelationships}{DataPackage}(x, ...)
}
\arguments{
\item{x}{A DataPackage object}

\item{...}{(Not yet used)}
}
\description{
Relationships of objects in a package are defined using the \code{'insertRelationship'} call and retrieved
using \code{getRetaionships}. These relationships are returned in a data frame with \code{'subject'}, \code{'predicate'}, \code{'objects'}
as the columns, ordered by "subject"
}
\section{Methods (by class)}{
\itemize{
\item \code{DataPackage}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

