% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSize}
\alias{getSize}
\title{Get the Count of Objects in the Package}
\usage{
getSize(x, ...)
}
\arguments{
\item{x}{DataPackage}

\item{...}{(not yet used)}
}
\value{
the number of object in the Package
}
\description{
Get the Count of Objects in the Package
}

