% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getSize}
\alias{getSize}
\alias{getSize,DataPackage-method}
\title{Get the Count of Objects in the Package}
\usage{
getSize(x, ...)

\S4method{getSize}{DataPackage}(x)
}
\arguments{
\item{x}{A DataPackage instance}

\item{...}{(not yet used)}
}
\value{
The number of object in the Package
}
\description{
Get the Count of Objects in the Package
}
\section{Methods (by class)}{
\itemize{
\item \code{DataPackage}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

