% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{getSize}
\alias{getSize}
\alias{getSize,DataPackage-method}
\title{Get the Count of Objects in the Package}
\usage{
getSize(x, ...)

\S4method{getSize}{DataPackage}(x)
}
\arguments{
\item{x}{A DataPackage instance}

\item{...}{(not yet used)}
}
\value{
The number of object in the Package
}
\description{
Get the Count of Objects in the Package
}
\examples{
dp <- new("DataPackage")
data <- charToRaw("1,2,3\\n4,5,6")
do <- new("DataObject", dataobj=data, format="text/csv", user="jsmith")
dp <- addMember(dp, do)
getSize(dp)
}
\seealso{
\code{\link{DataPackage-class}}
}

