% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{hasAccessRule}
\alias{hasAccessRule}
\alias{hasAccessRule,SystemMetadata,character-method}
\title{Determine if a particular access rules exists within SystemMetadata.}
\usage{
hasAccessRule(sysmeta, subject, ...)

\S4method{hasAccessRule}{SystemMetadata,character}(sysmeta, subject, permission)
}
\arguments{
\item{sysmeta}{the SystemMetadata instance to which to add the rules}

\item{subject}{the subject of the rule to be checked}

\item{...}{Additional arguments}

\item{permission}{the permission to be applied to subject if x is character}
}
\value{
A logical value: if TRUE the access rule was found, if FALSE it was not found.

boolean TRUE if the access rule exists already, FALSE otherwise
}
\description{
Each SystemMetadata document may contain a set of (subject, permission) tuples
that represent the access rules for its associated object. This method determines
whether a particular access rule already exists within the set.
}
\seealso{
\code{\link{SystemMetadata-class}{SystemMetadata}}{ class description.}
}

