% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{DataPackage}
\alias{DataPackage}
\alias{initialize,DataPackage-method}
\alias{initialize-DataPackage}
\title{Create a DataPackage object}
\usage{
DataPackage(...)

\S4method{initialize}{DataPackage}(.Object, packageId)
}
\arguments{
\item{...}{(Not yet used)}

\item{.Object}{The object being initialized}

\item{packageId}{The package id to assign to the package}
}
\description{
The DataPackage() method is a shortcut to creating a DataPackage object, as does
not allow specifying any options that the \code{\link[=initialize-DataPackage]{initialize}} method allows.

Initialize a DataPackage object
}
\examples{
# Create a DataPackage with undefined package id (to be set manually later)
pkg <- new("DataPackage")
# Alternatively, assign the package id when the DataPackage object is created
pkg <- new("DataPackage", "urn:uuid:4f953288-f593-49a1-adc2-5881f815e946")
}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

