% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemMetadata.R
\docType{methods}
\name{parseSystemMetadata}
\alias{parseSystemMetadata}
\alias{parseSystemMetadata,SystemMetadata,XMLInternalElementNode-method}
\title{Parse an external XML document and populate a SystemMetadata object with the parsed data}
\usage{
parseSystemMetadata(sysmeta, xml, ...)

\S4method{parseSystemMetadata}{SystemMetadata,XMLInternalElementNode}(sysmeta,
  xml, ...)
}
\arguments{
\item{sysmeta}{The \code{SystemMetadata} object}

\item{xml}{The XML representation of the capabilities, as an XMLInternalElementNode}

\item{...}{Additional arguments passed to other functions or methods}
}
\value{
the SystemMetadata object representing an object
}
\description{
Parse an XML representation of system metadata, and set the object slots of a SystemMetadata object 
the with obtained values.
}
\seealso{
\code{\link[=SystemMetadata-class]{SystemMetadata}}{ class description.}
}

