% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\name{removeMember}
\alias{removeMember}
\title{Remove the Specified Member from the Package}
\usage{
removeMember(x, identifier, ...)
}
\arguments{
\item{x}{a Datapackage object}

\item{identifier}{an identifier for a DataObject}
}
\description{
Given the identifier of a member of the data package, delete the DataObject
representation of the member.
}

