% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ResourceMap-class}
\alias{ResourceMap-class}
\title{ResourceMap provides methods to create, serialize and deserialize an OAI ORE resource map.}
\description{
The Open Archives Initiative Object Reuse and Exchange (OAI-ORE) defines standards for the description
and exchange of aggregrations of web resources, such as a DataPackage. A Resource Map describes the objects
in a DataPackage and the relationships between these objects.
}
\section{Slots}{

\describe{
\item{\code{relations}}{value of type \code{"data.frame"}, containing RDF triples representing the relationship between package objects}

\item{\code{world}}{a Redland RDF World object}

\item{\code{storage}}{a Redland RDF Storage object}

\item{\code{model}}{a Redland RDF Model object}

\item{\code{id}}{a unique identifier for a ResourceMap instance}
}}
\examples{
relations <- getRelationships(myDataPackage)
resMap <- new("ResourceMap", relations)
resMap <- createFromTriples(resMap, relations, getIdentifiers(myDataPackage))
serializeToRDF("ResourceMap", file="myResMap.rdf")
}
\author{
Peter Slaughter
}
\keyword{resourceMap}

