% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceMap.R
\docType{methods}
\name{initialize,ResourceMap-method}
\alias{initialize,ResourceMap-method}
\alias{ResourceMap-initialize}
\title{Initialize a ResourceMap object.}
\usage{
\S4method{initialize}{ResourceMap}(.Object, id = as.character(NA))
}
\arguments{
\item{.Object}{a ResourceMap object}

\item{id}{a unique identifier to identify this ResourceMap. This id will be used internally in the ResourceMap.}
}
\value{
the ResourceMap object
}
\description{
Create a ResourceMap object that contains relationships (RDF triples) of objects in the DataPackage.
}
\seealso{
\code{\link{ResourceMap-class}}
}
