% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{serializePackage}
\alias{serializePackage}
\alias{serializePackage,DataPackage-method}
\title{Create an OAI-ORE resource map from the package}
\usage{
serializePackage(.Object, ...)

\S4method{serializePackage}{DataPackage}(.Object, file, id = as.character(NA),
  syntaxName = "rdfxml", mimeType = "application/rdf+xml",
  namespaces = data.frame(namespace = character(), prefix = character(),
  stringsAsFactors = FALSE), syntaxURI = as.character(NA))
}
\arguments{
\item{.Object}{A DataPackage object}

\item{...}{Additional arguments}

\item{file}{The file to which the ResourceMap will be serialized}

\item{id}{A unique identifier for the serialization. If a value is not specified, one will be generated}

\item{syntaxName}{The name of the syntax to use for serialization - default is "rdfxml"}

\item{mimeType}{The mimetype of the serialized output - the default is "application/rdf+xml"}

\item{namespaces}{A data frame containing one or more namespaces and their associated prefix}

\item{syntaxURI}{URI of the serialization syntax}
}
\value{
A ResourceMap object
}
\description{
Datapackage relationships are serialized as a OAI-ORE resource map
and returned as a \code{\link{ResourceMap-class}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataPackage}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

