% Generated by roxygen2 (4.0.2): do not edit by hand
\name{serializeRDF}
\alias{serializeRDF}
\title{Serialize a ResouceMap.}
\usage{
serializeRDF(.Object, file, ...)
}
\arguments{
\item{.Object}{a ResourceMap}

\item{file}{the file to which the ResourceMap will be serialized}

\item{syntaxName}{name of the syntax to use for serialization - default is "rdfxml"}

\item{mimetype}{the mimetype of the serialized output - the default is "application/rdf+xml"}

\item{namespaces}{a data frame containing one or more namespaces and their associated prefix}

\item{resourceMapURI}{URI of the serialized syntax}
}
\value{
status of the serialization (non)
}
\description{
The Redland RDF library is used to serialize the ResourceMap RDF model
to a file as RDF/XML.
}

