% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ResourceMap.R
\docType{methods}
\name{serializeRDF}
\alias{serializeRDF}
\alias{serializeRDF,ResourceMap,character-method}
\title{Serialize a ResouceMap.}
\usage{
serializeRDF(.Object, file, ...)

\S4method{serializeRDF}{ResourceMap,character}(.Object, file,
  syntaxName = "rdfxml", mimeType = "application/rdf+xml",
  namespaces = data.frame(namespace = character(), prefix = character(),
  stringsAsFactors = FALSE), syntaxURI = as.character(NA))
}
\arguments{
\item{.Object}{a ResourceMap}

\item{file}{the file to which the ResourceMap will be serialized}

\item{...}{Additional parameters}

\item{syntaxName}{name of the syntax to use for serialization - default is "rdfxml"}

\item{mimeType}{the mimetype of the serialized output - the default is "application/rdf+xml"}

\item{namespaces}{a data frame containing one or more namespaces and their associated prefix}

\item{syntaxURI}{A URI of the serialized syntax}
}
\value{
status of the serialization (non)
}
\description{
The Redland RDF library is used to serialize the ResourceMap RDF model
to a file as RDF/XML.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = ResourceMap,file = character}: 
}}
\seealso{
\code{\link[=ResourceMap-class]{ResourceMap}}{ class description.}
}

