% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataPackage.R
\docType{methods}
\name{serializeToBagit}
\alias{serializeToBagit}
\alias{serializeToBagit,DataPackage-method}
\title{Serialize A DataPackage into a Bagit Archive File}
\usage{
serializeToBagit(.Object, ...)

\S4method{serializeToBagit}{DataPackage}(.Object)
}
\arguments{
\item{.Object}{A DataPackage object}

\item{...}{Additional arguments}
}
\value{
A zip file containing a Bagit archive.
}
\description{
The Bagit packaging format \url{https://tools.ietf.org/html/draft-kunze-bagit-08}
is used to prepare an archive file that contains the contents of a DataPackage.
}
\details{
A Bagit Archive File is created by copying each member of a DataPackge, and preparing
files that describe the files in the archive, including information about the size of the files
and a checksum for each file. These metadata files and the data files are then packaged into
a single zip file.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataPackage}: 
}}
\seealso{
\code{\link[=DataPackage-class]{DataPackage}}{ class description.}
}

