% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setPublicAccess}
\alias{setPublicAccess}
\alias{setPublicAccess,DataObject-method}
\title{Add a Rule to the AccessPolicy to make the object publicly readable}
\usage{
setPublicAccess(x, ...)

\S4method{setPublicAccess}{DataObject}(x)
}
\arguments{
\item{x}{DataObject}

\item{...}{(not yet used)}
}
\description{
To be called prior to creating the object in DataONE.  When called before
creating the object, adds a rule to the access policy that makes this object
publicly readable.  If called after creation, it will only change the system
metadata locally, and will not have any affect.
}
\section{Methods (by class)}{
\itemize{
\item \code{DataObject}: 
}}

