% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SystemMetadata-class}
\alias{SystemMetadata-class}
\title{A DataONE SystemMetadata object containing basic identification, ownership, access policy, replication policy, and related metadata.}
\description{
A class representing DataONE SystemMetadata, which is core information about objects stored in a repository
and needed to manage those objects across systems.  SystemMetadata contains basic identification, ownership,
access policy, replication policy, and related metadata.
}

