% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{validate-methods}
\alias{validate,SystemMetadata-method}
\alias{validate-methods}
\title{validate system metadata}
\usage{
validate(object, ...)

\S4method{validate}{SystemMetadata}(object, ...)
}
\arguments{
\item{object}{the instance to be validated}
}
\value{
logical, \code{TRUE} if the SystemMetadata object is valid, else a list of strings detailing errors
}
\description{
Validate an the system metadata object, ensuring that required fields are present and of the right type.
}
\author{
jones
}

